Add multitenancy v2
- create a shared schema
- migrate colleges to this shared schema
- combine all users across all tenants into one table in this shared
schema
- move foreign key references to these new tables
- add new superadmin role that allows for admin rights across all
tenants
- add new authorization rules that prevent users from logging into other
colleges (previously it was assumed they did not have accounts in other
colleges)

Resolves #802, #803, #804, #805, #806

+++ app/services/destroyers/destroyer.rb
@@ -3,6 +3,7 @@
 #
 # Base class for destroyer service objects. Provides default messages.
 class Destroyer
+  include ActiveModel::Model
   include Callable
 
   # Initialize a new Destroyer.
@@ -22,6 +23,7 @@ class Destroyer
   #   A results hash with a message to set in the flash and either `nil`
   #   or the object that was not destroyed
   def destroy
+    return error unless valid?
     if object.destroy
       success
     else
